agents/fisica/pyproject.toml
[project]
name = "physics-agent-adk"
version = "0.1.0"
description = "Physics Agent based on Google ADK hosted as an A2A server." 
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "a2a-samples",
    "click>=8.1.8",
    "google-adk>=0.0.3",
    "google-genai>=1.9.0",
    "python-dotenv>=1.1.0",
    "pypdf2>=3.0.0",
    "qdrant-client>=1.6.0",
    "transformers>=4.34.0",
    "torch>=2.0.0",
    "numpy>=1.24.0",
    "langchain>=0.0.312",
    "langchain-google-genai>=0.0.5",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv.sources]
a2a-samples = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"